% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/race.R
\name{race}
\alias{race}
\title{A race demography function.}
\usage{
race(t_number, end_year)
}
\arguments{
\item{t_number}{race table from acs package.}

\item{end_year}{survey year}
}
\description{
A race demography function.
}
\examples{
race()
}
\keyword{zipdemacs}
